Temporary workaround for uboot networking issue
If BMC network is plugged into "busy" switch and the
delay between uboot end and kernel start is short,
then the kernel will panic.  We didn't see this before
because it took a long time to decompress kernel and most
of our switches are not busy. We know if we wait at u-boot
prompt, then boot, then no kernel panic.  We also know if
BMC network is not plugged in, the no kernel panic.  This
moves uboot autoboot delay from 3s to 10s and if kernel
panics, watchdog will reboot.

Signed-off-by: Norman James <nkskjames@gmail.com>